Fix inlining of Metal Lock API
Declaring the functions inline in the header and defining them in the
source file was breaking the linker. Define and declare the functions in
the header and declare them extern inline in the source file.
